fix bash

+++ scripts/aws/ApplicationStop.sh
@@ -1,6 +1,15 @@
 #!/usr/bin/env bash
 
 cd ~/rian
-sudo docker-compose down
-sudo rm -rf ~/rian/*
+file="docker-compose.yml"
+if [ -f "$file" ]
+then
+	echo "$file found, close docker and remove existing files"
+	sudo docker-compose down
+	sudo rm -rf ~/rian/*
+else
+	echo "$file not found, continue download bundle"
+fi
+
+
